changed asset widget to not break our code
